3 qubit code and modules
I created a few modules to be able to use depending on what we are trying to use. We can continue to add to them as we create more functions.

I also fixed up the 3 qubit code so that it only has the function calls without the function itself in there because it would probably be easier for the user.
